Added French version of Dash (Fari Games) (#62)
closes #61

For this French translation, I've tried to use as much as possible the current French translation of Charge SRD, which is available in the same repository. Some paragraphs and terms are directly copied from it, and I'd like to thank the translators for their remarkable work.

Although, I've adopted a major change I want to explicit here. The English term _momentum_ was translated as _initiative_ in the Charge translation. Personally, I felt it didn't fit the idea of a pushing-forward force, something that was increasing the pace of the action as the story progresses. "Initiative" felt like something happening at the _beginning_ of the story, and not throughout it.

So I've picked the French word _impulsion_, because I think it reflects the idea I've expressed just above: something accelerating the rhythm, building up the… momentum.